refactor(genericserver): avoid huge functions
