+++ pkg/apis/core/install/roundtrip_test.go
@@ -1,5 +1,6 @@
 package install_test
 
+//nolint:godox // I don't understand the fuzzer yet, but I want to track this as a TODO.
 // TODO: Implement roundtrip tests for the core API group.
 //       This requires creating a fuzzer implementation.
 //       Reference: https://github.com/kubernetes/sample-apiserver/blob/master/pkg/apis/wardle/install/roundtrip_test.go
